+++ public/user-client/js/controllers/document/reviseController.js
@@ -16,6 +16,44 @@ app.controller("documentReviseController", function($scope, $rootScope, $filter,
     $scope.redirect = function(path){
         $location.url(path);
     };
+    
+    /**
+     * []
+     */
+    $scope.resetFile = function(){
+        $scope.status = "fail";
+        $scope.latest_revision.concerns.q14_filename = null;
+    };
+    
+    /**
+     * [upload file]
+     * @return {[type]} [file]
+     */
+    $scope.doUpload = function (files) {
+        $scope.status = "uploading";
+        upload({
+          url: config.getUploadEndpoint() + $scope.latest_revision.concerns.concern_id,
+          method: 'POST',
+          headers: {
+              'Authorization': 'Bearer ' + $authenticationService.getToken(),
+              'X-DocumentId' : $documentService.getId()
+            },
+          data: {
+            filename: files[0], // a jqLite type="file" element, upload() will extract all the files from the input and put them into the FormData object before sending.
+          }
+        }).then(
+          function (response) {
+            $scope.status = "success";
+            $scope.latest_revision.concerns.q14_file = true;
+            $scope.latest_revision.concerns.q14_filename = files[0].name;
+            $scope.latest_revision.concerns.q14_filepath = response.data;
+          },
+          function (response) {
+            $scope.status = "fail";
+            $window.alert($filter('translate')('ALERT_UPLOAD_FILE_FAILED'));
+          }
+        );
+    }
 
     /**
      * [toggleConcernHistory description]
@@ -251,6 +289,7 @@ app.controller("documentReviseController", function($scope, $rootScope, $filter,
             $scope.editDocumentForm.q11_2_value.$pristine = false;
             $scope.editDocumentForm.q12_value.$pristine = false;
             $scope.editDocumentForm.q13_value.$pristine = false;
+            $scope.editDocumentForm.q14_value.$pristine = false;
 
             // Conerns (explanations)
             if($scope.editDocumentForm.q01_explanation){
@@ -294,6 +333,9 @@ app.controller("documentReviseController", function($scope, $rootScope, $filter,
             }
             if($scope.editDocumentForm.q13_explanation){
                 $scope.editDocumentForm.q13_explanation.$pristine = false;
+            }
+            if($scope.editDocumentForm.q14_explanation){
+                $scope.editDocumentForm.q14_explanation.$pristine = false;
             }
 
             $window.alert($filter('translate')('ALERT_SUBMIT_DOCUMENT_FAILED'));