+++ reminder.js
@@ -372,6 +372,14 @@ async.waterfall([
                         concern.q13_label = "badge-success";
                         concern.q13_sign = "no";
                     }
+                    
+                    if(concern.q14_value){
+                        concern.q14_label = "badge-danger";
+                        concern.q14_sign = "yes";
+                    } else {
+                        concern.q14_label = "badge-success";
+                        concern.q14_sign = "no";
+                    }
 
                     // Notify each committee member
                     async.eachOfSeries(members, function (member, key, callback) {