+++ sql/migration/migrate_100_to_141.sql
@@ -1,7 +1,9 @@
-
-TODO(specki): add file column to description
-
+-- Add to Concerns
 ALTER TABLE Concerns ADD q14_value boolean default null;
 ALTER TABLE Concerns ADD q14_explanation text default null;
 ALTER TABLE Concerns ADD q14_filename text default null;
 ALTER TABLE Concerns ADD q14_filepath text default null;
+
+-- Add to Comments
+
+ALTER TABLE Comments ADD q14_comment text default null;