+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Version 1.4.1
  - Added Email Notification to all Members when Review is started (and not pending anymore)
+ - Added q14 (Custom Files)
 
 # Version 1.4.0
  - Added Reminder Email to be sent out after 7 days of inactivity with still pending reviews.
@@ -12,3 +13,7 @@
 
 # Version 1.0
  - Initial App
+
+
+# TODO
+ - q14 in email templates